# flattenLet: only deshadow when needed