Before, we called `freshenTm` blindly on the entire (nested) let-expression. But we can, and should, deshadow in a more piecemail fashion; i.e. only deshadow the inner let-expressions that need it.

Before:
```
benchmarking normalization of examples/Reducer.hs
time                 2.444 s    (2.391 s .. 2.541 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.416 s    (2.405 s .. 2.434 s)
std dev              16.70 ms   (1.628 ms .. 21.20 ms)
variance introduced by outliers: 19% (moderately inflated)
```

After:
```
benchmarking normalization of examples/Reducer.hs
time                 1.743 s    (1.700 s .. 1.770 s)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.725 s    (1.714 s .. 1.732 s)
std dev              10.57 ms   (4.714 ms .. 13.92 ms)
variance introduced by outliers: 19% (moderately inflated)
```